odoo_contacts/requirements.txt
aiofiles==23.2.1
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
asyncpgsa==0.27.1
certifi==2023.11.17
click==8.1.7
croniter==1.4.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-filter==1.0.0
fastapi-filters==0.2.9
fastapi-utilities==0.1.3.1
fastapi_amis_admin==0.6.9
fastapi_scheduler==0.0.15
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.1.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.1
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydentic==0.0.1.dev3
pytest==7.4.3
pytest-asyncio==0.23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-stdnum==1.19
pytz==2023.3.post1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy_database==0.1.1
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.9.0
tzlocal==5.2
uvicorn==0.24.0.post1
